docs: clean up documentation and comments

+++ src/lucky_honeypot/signals.cr
@@ -14,14 +14,13 @@ struct LuckyHoneypot::Signals
   @[JSON::Field(key: "f")]
   getter? focus : Bool
 
-  # Calculates the likelyhood of a human (0 being a bot, 1 being a human). Any
-  # value above 0 is already a good chance.
+  # Calculates the likelihood of a human (0 being a bot, 1 being a human).
   def human_rating : Float64
     [mouse?, touch?, scroll?, keyboard?, focus?].count(&.itself) / 5.0
   end
 
   # Convenience method to calculate the human rating for a class method.
-  def self.human_rating(json : String) : Float
+  def self.human_rating(json : String) : Float64
     from_json(json).human_rating
   end
 end
